+++ lib/Service/SettingsService.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace OCA\Files_Confidential\Service;
 
 use OCA\Files_Confidential\Model\ClassificationLabel;
@@ -8,12 +10,10 @@ use Psr\Log\LoggerInterface;
 use Safe\Exceptions\JsonException;
 
 class SettingsService {
-	private IConfig $config;
-	private LoggerInterface $logger;
-
-	public function __construct(IConfig $config, LoggerInterface $logger) {
-		$this->config = $config;
-		$this->logger = $logger;
+	public function __construct(
+		private IConfig $config,
+		private LoggerInterface $logger
+	) {
 	}
 
 	/**
@@ -21,7 +21,10 @@ class SettingsService {
 	 */
 	public function getClassificationLabels(): array {
 		try {
-			/** @var array $labelsRaw */
+			/**
+			 * @var array $labelsRaw
+			 * @psalm-suppress DeprecatedMethod
+			 */
 			$labelsRaw = \Safe\json_decode($this->config->getAppValue('files_confidential', 'labels', '[]'), true);
 		} catch (JsonException $e) {
 			$this->logger->warning('Could not load labels setting', ['exception' => $e]);
@@ -56,6 +59,7 @@ class SettingsService {
 			$this->logger->warning('Could not store labels setting', ['exception' => $e]);
 			throw $e;
 		}
+		/** @psalm-suppress DeprecatedMethod */
 		$this->config->setAppValue('files_confidential', 'labels', $json);
 	}
 }