+++ lib/Settings/AdminSettings.php
@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /*
  * Copyright (c) 2021-2022 The Recognize contributors.
  * This file is licensed under the Affero General Public License version 3 or later. See the COPYING file.
@@ -9,19 +12,15 @@ namespace OCA\Files_Confidential\Settings;
 use OCA\Files_Confidential\Service\MatcherService;
 use OCA\Files_Confidential\Service\SettingsService;
 use OCP\AppFramework\Http\TemplateResponse;
-use OCP\IInitialStateService;
+use OCP\AppFramework\Services\IInitialState;
 use OCP\Settings\ISettings;
 
 class AdminSettings implements ISettings {
-	private SettingsService $settingsService;
-	private IInitialStateService $initialState;
-
-	private MatcherService $matcherService;
-
-	public function __construct(SettingsService $settingsService, IInitialStateService $initialState, MatcherService $matcherService) {
-		$this->settingsService = $settingsService;
-		$this->initialState = $initialState;
-		$this->matcherService = $matcherService;
+	public function __construct(
+		private SettingsService $settingsService,
+		private IInitialState $initialState,
+		private MatcherService $matcherService
+	) {
 	}
 
 	/**
@@ -30,21 +29,21 @@ class AdminSettings implements ISettings {
 	public function getForm(): TemplateResponse {
 		$labels = $this->settingsService->getClassificationLabels();
 		$labels = array_map(fn ($label) => $label->toArray(), $labels);
-		$this->initialState->provideInitialState('files_confidential', 'labels', $labels);
-		$this->initialState->provideInitialState('files_confidential', 'searchExpressions', $this->matcherService->expressions);
+		$this->initialState->provideInitialState('labels', $labels);
+		$this->initialState->provideInitialState('searchExpressions', $this->matcherService->expressions);
 
 		return new TemplateResponse('files_confidential', 'admin');
 	}
 
 	/**
-	 * @return string the section ID, e.g. 'sharing'
+	 * @inheritdoc
 	 */
 	public function getSection(): string {
 		return 'files_confidential';
 	}
 
 	/**
-	 * @return int whether the form should be rather on the top or bottom of the admin section. The forms are arranged in ascending order of the priority values. It is required to return a value between 0 and 100.
+	 * @inheritdoc
 	 */
 	public function getPriority(): int {
 		return 1;