+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2024-03-06
+
+Maintenance update.
+
+### Added
+
+- Added support for Nextcloud 29
+
+### Fixed
+
+- Fixed upload of policy does not reload schemas in admin settings
+- Fixes other minor visual issues on the admin settings page
+
+### Updated
+
+- Updated npm packages (security, Nextcloud Vue 8)
+
 ## [2.0.1] - 2023-11-08
 
 ### Fixed